+++ packages/terra-application-utility/src/ApplicationHeaderUtility.jsx
@@ -30,6 +30,14 @@ const propTypes = {
    * The accessory element to be displayed next to the title.
    */
   accessory: PropTypes.element,
+  /**
+   * The role attribute to set on the menu.
+   */
+  menuRole: PropTypes.string,
+};
+
+const defaultProps = {
+  menuRole: 'navigation',
 };
 
 const ApplicationHeaderUtility = ({
@@ -39,6 +47,7 @@ const ApplicationHeaderUtility = ({
   initialSelectedKey,
   title,
   accessory,
+  menuRole,
   ...customProps
 }) => (
   <UtilityButton
@@ -49,10 +58,12 @@ const ApplicationHeaderUtility = ({
     initialSelectedKey={initialSelectedKey}
     title={title}
     accessory={accessory}
+    menuRole={menuRole}
     variant={Utils.VARIANTS.HEADER}
   />
 );
 
 ApplicationHeaderUtility.propTypes = propTypes;
+ApplicationHeaderUtility.defaultProps = defaultProps;
 
 export default ApplicationHeaderUtility;