+++ packages/terra-application-utility/src/ApplicationMenuUtility.jsx
@@ -30,15 +30,24 @@ const propTypes = {
    * The accessory element to be displayed next to the title.
    */
   accessory: PropTypes.element,
+  /**
+   * The role attribute to set on the menu.
+   */
+  menuRole: PropTypes.string,
 };
 
-const ApplicationHeaderUtility = ({
+const defaultProps = {
+  menuRole: 'navigation',
+};
+
+const ApplicationMenuUtility = ({
   menuItems,
   onChange,
   onDisclose,
   initialSelectedKey,
   title,
   accessory,
+  menuRole,
   ...customProps
 }) => (
   <UtilityButton
@@ -49,10 +58,12 @@ const ApplicationHeaderUtility = ({
     initialSelectedKey={initialSelectedKey}
     title={title}
     accessory={accessory}
+    menuRole={menuRole}
     variant={Utils.VARIANTS.MENU}
   />
 );
 
-ApplicationHeaderUtility.propTypes = propTypes;
+ApplicationMenuUtility.propTypes = propTypes;
+ApplicationMenuUtility.defaultProps = defaultProps;
 
-export default ApplicationHeaderUtility;
+export default ApplicationMenuUtility;